Rename TreeNode format to "Simple" and "Delta"
