+++ enterprise/index/src/main/resources/META-INF/services/org.neo4j.index.internal.gbptree.TreeNodeFactory
@@ -1,1 +1,1 @@
-org.neo4j.index.internal.gbptree.TreeNodeV3Factory
+org.neo4j.index.internal.gbptree.TreeNodeDeltaFactory
